+++ crates/cheatnet/tests/cheatcodes/prank.rs
@@ -52,7 +52,7 @@ fn prank_in_constructor() {
 
     state.start_prank(precalculated_address, ContractAddress::from(123_u128));
 
-    let contract_address = state.deploy(&class_hash, &[]).unwrap();
+    let contract_address = state.deploy(&class_hash, &[]).unwrap().contract_address;
 
     assert_eq!(precalculated_address, contract_address);