+++ crates/conversions/src/field_element.rs
@@ -1,6 +1,7 @@
 use super::StarknetConversions;
 use cairo_felt::Felt252;
 use starknet::core::types::FieldElement;
+use starknet_api::core::Nonce;
 use starknet_api::{
     core::{ClassHash, ContractAddress},
     hash::{StarkFelt, StarkHash},
@@ -33,5 +34,9 @@ impl StarknetConversions for FieldElement {
 
     fn to_short_string(&self) -> String {
         self.to_felt252().to_short_string()
+    }
+
+    fn to_nonce(&self) -> Nonce {
+        self.to_felt252().to_nonce()
     }
 }