+++ crates/forge/src/running.rs
@@ -1,8 +1,17 @@
 use std::collections::HashMap;
+use std::sync::Arc;
 
 use anyhow::Result;
 use cairo_felt::Felt252;
+use blockifier::execution::entry_point::{
+    CallEntryPoint, CallType, EntryPointExecutionContext, ExecutionResources,
+};
+use blockifier::execution::execution_utils::ReadOnlySegments;
+use blockifier::execution::syscalls::hint_processor::SyscallHintProcessor;
+use blockifier::state::cached_state::{CachedState, GlobalContractCache};
 use cairo_vm::serde::deserialize_program::HintParams;
+use cairo_vm::types::relocatable::Relocatable;
+use cheatnet::constants::{build_block_context, build_testing_state, build_transaction_context};
 use cheatnet::CheatnetState;
 use itertools::chain;
 
@@ -13,8 +22,14 @@ use cairo_lang_runner::{Arg, CairoHintProcessor as CoreCairoHintProcessor, Runne
 use cairo_lang_runner::{SierraCasmRunner, StarknetState};
 use cairo_vm::vm::runners::cairo_runner::RunResources;
 use camino::Utf8PathBuf;
-use cheatnet::constants::build_testing_state;
-use cheatnet::state::StateReaderProxy;
+use cheatnet::state::ExtendedStateReader;
+use starknet::core::utils::get_selector_from_name;
+use starknet_api::core::PatriciaKey;
+use starknet_api::core::{ContractAddress, EntryPointSelector};
+use starknet_api::deprecated_contract_class::EntryPointType;
+use starknet_api::hash::StarkHash;
+use starknet_api::patricia_key;
+use starknet_api::transaction::Calldata;
 use test_collector::TestCase;
 
 use crate::cheatcodes_hint_processor::CairoHintProcessor;
@@ -59,6 +74,7 @@ pub(crate) fn run_from_test_case(
     } else {
         Some(usize::MAX)
     };
+
     let func = runner.find_function(case.name.as_str())?;
     let initial_gas = runner.get_initial_available_gas(func, available_gas)?;
 
@@ -72,18 +88,56 @@ pub(crate) fn run_from_test_case(
         footer.iter()
     );
     let (hints_dict, string_to_hint) = build_hints_dict(instructions.clone());
-    let core_cairo_hint_processor = CoreCairoHintProcessor {
-        runner: Some(runner),
-        starknet_state: StarknetState::default(),
-        string_to_hint,
-        run_resources: RunResources::default(),
+
+    // Losely inspired by crates/cheatnet/src/execution/cairo1_execution::execute_entry_point_call_cairo1
+    let block_context = build_block_context();
+    let account_context = build_transaction_context();
+    let mut context = EntryPointExecutionContext::new(
+        block_context.clone(),
+        account_context,
+        block_context.invoke_tx_max_n_steps.try_into().unwrap(),
+    );
+    let test_selector = get_selector_from_name("TEST_CONTRACT_SELECTOR").unwrap();
+    let entry_point_selector = EntryPointSelector(StarkHash::new(test_selector.to_bytes_be())?);
+    let entry_point = CallEntryPoint {
+        class_hash: None,
+        code_address: Some(ContractAddress::from(0_u8)),
+        entry_point_type: EntryPointType::External,
+        entry_point_selector,
+        calldata: Calldata(Arc::new(vec![])),
+        storage_address: ContractAddress(patricia_key!("0x0")),
+        caller_address: ContractAddress::default(),
+        call_type: CallType::Call,
+        initial_gas: u64::MAX,
     };
+
+    let mut blockifier_state = CachedState::new(
+        build_testing_state(predeployed_contracts),
+        GlobalContractCache::default(),
+    );
+    let mut execution_resources = ExecutionResources::default();
+    let syscall_handler = SyscallHintProcessor::new(
+        &mut blockifier_state,
+        &mut execution_resources,
+        &mut context,
+        // This segment is created by SierraCasmRunner
+        Relocatable {
+            segment_index: 10,
+            offset: 0,
+        },
+        entry_point,
+        &string_to_hint,
+        ReadOnlySegments::default(),
+    );
     let mut cairo_hint_processor = CairoHintProcessor {
-        original_cairo_hint_processor: core_cairo_hint_processor,
+        blockifier_syscall_handler: syscall_handler,
         contracts,
-        cheatnet_state: CheatnetState::new(StateReaderProxy(Box::new(build_testing_state(
-            predeployed_contracts,
-        )))),
+        cheatnet_state: CheatnetState::new(ExtendedStateReader {
+            dict_state_reader: build_testing_state(predeployed_contracts),
+            fork_state_reader: None,
+        }),
+        hints: &string_to_hint,
+        run_resources: RunResources::default(),
     };
 
     match runner.run_function(