+++ crates/forge/tests/integration/syscalls.rs
@@ -131,28 +131,6 @@ fn library_call_syscall() {
 }
 
 #[test]
-fn test_call_syscall_fail_in_test_fn() {
-    let test = test_case!(indoc!(
-        r#"
-        use starknet::{ get_block_timestamp };
-        #[test]
-        fn test_execute_disallowed_syscall() {
-            get_block_timestamp();
-        }
-    "#
-    ));
-
-    let result = run_test_case(&test);
-
-    assert_case_output_contains!(
-        result,
-        "test_execute_disallowed_syscall",
-        "starknet syscalls (other than CallContract and Keccak) cannot be used in tests"
-    );
-    assert_failed!(result);
-}
-
-#[test]
 fn test_keccak_syscall() {
     let test = test_case!(indoc!(
         r#"
@@ -199,11 +177,7 @@ fn test_keccak_syscall_too_small_input() {
 
     let result = run_test_case(&test);
 
-    assert_case_output_contains!(
-        result,
-        "test_execute_cairo_keccak",
-        "Invalid keccak input size"
-    );
+    assert_case_output_contains!(result, "test_execute_cairo_keccak", "Invalid input length");
 
     assert_failed!(result);
 }